refactor(Utils): Create sum file
