+++ src/days/day01/factors-sum.js
@@ -1,28 +1,4 @@
-export function find2FactorsMakingUpSum(entries, searchValue) {
-    var left = 0;
-    var right = entries.length - 1;
-
-    var leftValue, rightValue, nextLeftValue;
-
-    while (left <= right) {
-        leftValue = entries[left];
-        rightValue = entries[right];
-
-        nextLeftValue = entries[left + 1];
-
-        if (leftValue + rightValue === searchValue) {
-            return [leftValue, rightValue];
-        }
-
-        if (nextLeftValue + rightValue <= searchValue) {
-            left = left + 1;
-        } else {
-            right = right - 1;
-        }
-    }
-
-    return;
-}
+import { find2FactorsMakingUpSum } from "../../utils/sum";
 
 export function find3FactorsMakingUp2020(entries) {
     const searchValue = 2020;
